+++ lib/overcommit/hook/pre_commit/json_syntax.rb
@@ -8,7 +8,7 @@ module Overcommit::Hook::PreCommit
 
       applicable_files.each do |file|
         begin
-          File.open(file) { |io| JSON.load(io) }
+          JSON.parse(IO.read(file))
         rescue JSON::ParserError => e
           output << "#{e.message} parsing #{file}"
         end